requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
python-decouple
dj-config-url